tools/requirements.txt
google-genai>=1.52.0
Pillow>=10.0.0
